Refactor the query/users struct, make user and stage more cleaner
